Changes in the code
