+++ Assignment/product.rb
@@ -1,13 +1,11 @@
 require 'yaml'
 
 class Product
-
   def product_list(data)
-    data.each do |data|
-      puts data.values
+    data.each do |data| puts data.values
     end
   end
-  
+
   def product_details
     products = YAML.load_file('products_data.yaml')
     product_list(products)
